Refactor scripts functionality to generate onload through domUpdates
