+++ src/domUpdates.js
@@ -1,0 +1,35 @@
+let domUpdates = {
+  // GENERATE A USER ON LOAD
+  generateUser(currentUser) {
+    let firstName = currentUser.name.split(" ")[0];
+    let welcomeMsg = `
+      <div class="welcome-msg">
+        <h1>Welcome ${firstName}!</h1>
+      </div>`;
+    document.querySelector(".banner-image").insertAdjacentHTML("afterbegin",
+      welcomeMsg);
+  },
+
+  // Create Recipe Cards
+
+  displayCard(recipeInfo, shortRecipeName) {
+    let main = document.querySelector("main");
+    let cardHtml = `
+      <div class="recipe-card" id=${recipeInfo.id}>
+        <h3 maxlength="40">${shortRecipeName}</h3>
+        <div class="card-photo-container">
+          <img src=${recipeInfo.image} class="card-photo-preview" alt="${recipeInfo.name} recipe" title="${recipeInfo.name} recipe">
+          <div class="text">
+            <div>Click for Instructions</div>
+          </div>
+        </div>
+        <h4>${recipeInfo.tags[0]}</h4>
+        <img src="./images/apple-logo-outline.png" alt="unfilled apple icon" class="card-apple-icon">
+      </div>`
+    main.insertAdjacentHTML("beforeend", cardHtml);
+  },
+
+
+}
+
+export default domUpdates;
